+++ src/main/java/matteroverdrive/compat/MatterOverdriveCompat.java
@@ -7,6 +7,7 @@ import java.util.ArrayList;
 
 import matteroverdrive.compat.modules.CompatTConstruct;
 import matteroverdrive.compat.modules.computercraft.CompatComputerCraft;
+import matteroverdrive.compat.modules.groovyscript.CompatGroovyScript;
 import matteroverdrive.compat.modules.sd.CompatSD;
 import matteroverdrive.compat.modules.sd.CompatSDClient;
 import matteroverdrive.compat.modules.tan.CompatTAN;
@@ -33,6 +34,7 @@ public class MatterOverdriveCompat {
 		register(CompatComputerCraft.class);
 		register(CompatTAN.class);
 		register(CompatSD.class);
+		register(CompatGroovyScript.class);
 	}
 
 	// Add additional modules that need to run ONLY on the Client